+++ packages/cli/src/util/projects/find-project-root.ts
@@ -6,6 +6,7 @@ import {
   isExperimentalServicesEnabled,
   tryDetectServices,
 } from './detect-services';
+import { isVercelTomlEnabled } from '../is-vercel-toml-enabled';
 import output from '../../output-manager';
 
 /**
@@ -42,9 +43,11 @@ export async function findProjectRoot(
 
     const hasVercelDir = await pathExists(join(dir, '.vercel'));
     const hasVercelJson = await pathExists(join(dir, 'vercel.json'));
+    const hasVercelToml =
+      isVercelTomlEnabled() && (await pathExists(join(dir, 'vercel.toml')));
     const hasGit = await pathExists(join(dir, '.git'));
 
-    if (hasVercelDir || hasVercelJson || hasGit) {
+    if (hasVercelDir || hasVercelJson || hasVercelToml || hasGit) {
       return dir;
     }