+++ packages/cli/src/util/telemetry/commands/teams/index.ts
@@ -42,6 +42,15 @@ export class TeamsTelemetryClient
     }
   }
 
+  trackCliSubcommandRequest(actual?: string) {
+    if (actual) {
+      this.trackCliSubcommand({
+        subcommand: 'request',
+        value: actual,
+      });
+    }
+  }
+
   trackCliSubcommandMembers(actual?: string) {
     if (actual) {
       this.trackCliSubcommand({